% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_clip.r
\name{df_to_clip}
\alias{df_to_clip}
\title{Copy a data.frame to the clipboard}
\usage{
df_to_clip(df, digits = 4)
}
\arguments{
\item{df}{A data.frame}

\item{digits}{number of digits for rounding numerical columns}
}
\value{
the data.frame is copied to the clipboard allowing for easy pasting into excel (and subsequently to ppt, etc.)
}
\description{
Copy a data.frame to the clipboard
}
