% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.bin.int.HelperFunctions.r
\name{Discord_Go}
\alias{Discord_Go}
\title{Discord_Go}
\usage{
Discord_Go(go, n, fgo, N, prob, Weights = NA)
}
\arguments{
\item{go}{go}

\item{n}{n}

\item{fgo}{fgo}

\item{N}{N}

\item{prob}{prob}

\item{Weights}{weights}
}
\value{
These functions are used by other functions
}
\description{
Discord_Go
}
