% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.bin.int.HelperFunctions.r
\name{Discord_Grey}
\alias{Discord_Grey}
\title{Discord_Grey}
\usage{
Discord_Grey(low, up, n, fgo, fnogo, N, prob, Weights = NA)
}
\arguments{
\item{low}{low}

\item{up}{up}

\item{n}{n}

\item{fgo}{fgo}

\item{fnogo}{gnogo}

\item{N}{N}

\item{prob}{prob}

\item{Weights}{weights}
}
\value{
These functions are used by other functions
}
\description{
Discord_Grey
}
