% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.bin.int.HelperFunctions.r
\name{Discord_NoGo}
\alias{Discord_NoGo}
\title{Discord_NoGo}
\usage{
Discord_NoGo(nogo, n, fnogo, N, prob, Weights = NA)
}
\arguments{
\item{nogo}{nogo}

\item{n}{n}

\item{fnogo}{fnogo}

\item{N}{N}

\item{prob}{prob}

\item{Weights}{Weights}
}
\value{
These functions are used by other functions
}
\description{
Discord_NoGo
}
