% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stepribbon.r
\name{geom_stepribbon}
\alias{geom_stepribbon}
\title{geom_stepribbon
\code{geom_stepribbon} is an extension of the \code{geom_ribbon}, and
is optimized for Kaplan-Meier plots with pointwise confidence intervals
or a confidence band.}
\usage{
geom_stepribbon(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  kmplot = FALSE,
  ...
)
}
\arguments{
\item{mapping}{see source reference}

\item{data}{see source reference}

\item{stat}{see source reference}

\item{position}{see source reference}

\item{na.rm}{see source reference}

\item{show.legend}{see source reference}

\item{inherit.aes}{see source reference}

\item{kmplot}{If \code{TRUE}, missing values are replaced by the previous
values. This option is needed to make Kaplan-Meier plots if the last
observation has event, in which case the upper and lower values of the
last observation are missing. This processing is optimized for results
from the survfit function.}

\item{...}{see source reference}
}
\value{
geom_stepribbon is an analog to geom_ribbon using step function.
}
\description{
Step ribbon plots.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "ribbon")}
}

\seealso{
\code{\link[ggplot2:geom_ribbon]{geom_ribbon}} \code{geom_stepribbon}
  inherits from \code{geom_ribbon}.
}
