% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.normal.ss.post.r
\name{get.normal.ss.post}
\alias{get.normal.ss.post}
\title{Get normal single sample posterior parameters}
\usage{
get.normal.ss.post(
  mu.0 = 0,
  n.0 = 10,
  alpha.0 = 0.25,
  beta.0 = 1,
  xbar = 4,
  s = 3,
  n = 15,
  group = "Placebo"
)
}
\arguments{
\item{mu.0}{prior mean}

\item{n.0}{prior effective sample size}

\item{alpha.0}{prior alpha parameter}

\item{beta.0}{prior beta parameter}

\item{xbar}{observed sampled mean}

\item{s}{observed sample standard deviation}

\item{n}{sample size}

\item{group}{text string for group label}
}
\value{
Returns normal single sample posterior parameters
}
\description{
Get normal single sample posterior parameters
}
\examples{
my.normal.ss.post <- get.normal.ss.post(mu.0=0, n.0=10, alpha.0=.25,
beta.0=1, xbar=4, s=3, n=15, group="Placebo")
my.normal.ss.post
}
\author{
Greg Cicconetti
}
