% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ss.bin.int.df.r
\name{get.ss.bin.int.df}
\alias{get.ss.bin.int.df}
\title{Get single sample binary interim data.frame}
\usage{
get.ss.bin.int.df(
  ss.bin.studyend.GNG = get.ss.bin.studyend.GNG(a.trt = 1, b.trt = 1, n.trt = 50, x.trt =
    9, Delta.lrv = 0.62, Delta.tv = 0.7, tau.tv = 0.1, tau.lrv = 0.8, tau.ng = 0.65),
  goThreshold = 0.8,
  nogoThreshold = 0.8,
  include_nogo = FALSE
)
}
\arguments{
\item{ss.bin.studyend.GNG}{The output from get.ss.bin.studyend.GNG}

\item{goThreshold}{predictive posterior probability of a final Go threshold for an interim Go}

\item{nogoThreshold}{predictive posterior probability of a final No Go threshold for an interim NoGo}

\item{include_nogo}{logical}
}
\value{
A dataframe holding interim pred prob of Go and No-Go along with interim decision
}
\description{
get.ss.bin.interim.df
}
