% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ss.bin.int.GNG.r
\name{get.ss.bin.int.GNG}
\alias{get.ss.bin.int.GNG}
\title{Get single sample binary interim Go-NoGo}
\usage{
get.ss.bin.int.GNG(ss.bin.int.df, Interims = 20, ss.bin.studyend.GNG)
}
\arguments{
\item{ss.bin.int.df}{output from get.ss.bin.interim.df}

\item{Interims}{vector of sample sizes at which interims will occur}

\item{ss.bin.studyend.GNG}{output from call to get.ss.bin.studyend.GNG}
}
\value{
A tibble is returned holding the min and max needed for Go and No-Go resp.
}
\description{
Get single sample binary interim Go-NoGo
}
\examples{
{
my.ss.bin.int.df <- get.ss.bin.int.df(ss.bin.studyend.GNG =
get.ss.bin.studyend.GNG(a.trt = 1, b.trt = 1, n.trt = 40, x.trt = 9,
Delta.lrv = .2, Delta.tv = .35, tau.tv = 0.10, tau.lrv = .80, tau.ng = .65),
goThreshold = .8, nogoThreshold = 1.2, include_nogo = TRUE)
my.ss.bin.studyend.GNG <-  get.ss.bin.studyend.GNG(a.trt = 1, b.trt = 1, n.trt = 40, x.trt = 9,
                                                   Delta.lrv = .2, Delta.tv = .35,
                                                   tau.tv = 0.10, tau.lrv = .80, tau.ng = .65)
   my.ss.bin.int.GNG <- get.ss.bin.int.GNG(
   ss.bin.int.df = my.ss.bin.int.df,
   Interims = 20,
   ss.bin.studyend.GNG = my.ss.bin.studyend.GNG)
 }

}
