% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ss.bin.ssize.oc.df.r
\name{get.ss.bin.ssize.oc.df}
\alias{get.ss.bin.ssize.oc.df}
\title{Get single sample binary sample size operating charactersitics data.frame}
\usage{
get.ss.bin.ssize.oc.df(
  a.trt = 1,
  b.trt = 1,
  n.trt = 40,
  Delta.lrv = 0.35,
  Delta.tv = 0.35,
  Delta.user = 0.4,
  tau.tv = 0.1,
  tau.lrv = 0.8,
  tau.ng = 0.65,
  SS.OC.N.LB = 20,
  SS.OC.N.UB = 60
)
}
\arguments{
\item{a.trt}{prior alpha parameter}

\item{b.trt}{prior beta parameter}

\item{n.trt}{observed sample size}

\item{Delta.lrv}{TPP Lower Reference Value aka Min TPP}

\item{Delta.tv}{TPP Target Value aka Base TPP}

\item{Delta.user}{User-specified rate}

\item{tau.tv}{threshold associated with Base TPP}

\item{tau.lrv}{threshold associated with Min TPP}

\item{tau.ng}{threshold associated with No-Go}

\item{SS.OC.N.LB}{Sample size OC Curve's Lower sample size bound}

\item{SS.OC.N.UB}{Sample size OC Curve's Upper sample size bound}
}
\value{
A data.frame is returned
}
\description{
Get single sample binary sample size operating charactersitics data.frame
}
\examples{
my.ss.bin.ssize.oc.df <- get.ss.bin.ssize.oc.df()
head(my.ss.bin.ssize.oc.df)
}
