% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ss.bin.studyend.GNG.r
\name{get.ss.bin.studyend.GNG}
\alias{get.ss.bin.studyend.GNG}
\title{Get single sample binary study end GNG criteria}
\usage{
get.ss.bin.studyend.GNG(
  a.trt = 1,
  b.trt = 1,
  n.trt = 40,
  x.trt = 9,
  Delta.lrv = 0.2,
  Delta.tv = 0.35,
  tau.tv = 0.1,
  tau.lrv = 0.8,
  tau.ng = 0.65
)
}
\arguments{
\item{a.trt}{prior alpha parameter}

\item{b.trt}{prior beta parameter}

\item{n.trt}{observed sample size}

\item{x.trt}{observed number of responders}

\item{Delta.lrv}{TPP Lower Reference Value aka Min TPP}

\item{Delta.tv}{TPP Target Value aka Base TPP}

\item{tau.tv}{threshold associated with Base TPP}

\item{tau.lrv}{threshold associated with Min TPP}

\item{tau.ng}{threshold associated with No-Go}
}
\value{
Returns a list of data.frames holding what is needed from data to acheive Go/No-Go
}
\description{
Get single sample binary study end GNG criteria
}
\examples{
get.ss.bin.studyend.GNG()
}
