% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ss.bin.trt.oc.df.r
\name{get.ss.bin.trt.oc.df}
\alias{get.ss.bin.trt.oc.df}
\title{Get single sample binary treatment OC data.frame}
\usage{
get.ss.bin.trt.oc.df(
  a.trt = 1,
  b.trt = 1,
  n.trt = 40,
  Delta.tv = 0.35,
  Delta.lrv = 0.2,
  tau.tv = 0.1,
  tau.lrv = 0.8,
  tau.ng = 0.65
)
}
\arguments{
\item{a.trt}{prior alpha parameter}

\item{b.trt}{prior beta parameter}

\item{n.trt}{observed sample size}

\item{Delta.tv}{TPP Target Value aka Base TPP}

\item{Delta.lrv}{TPP Lower Reference Value aka Min TPP}

\item{tau.tv}{threshold associated with Base TPP}

\item{tau.lrv}{threshold associated with Min TPP}

\item{tau.ng}{threshold associated with No-Go}
}
\value{
A data.frame is returned.
}
\description{
Get single sample binary treatment OC data.frame
}
\examples{
my.ss.bin.trt.oc.df <- get.ss.bin.trt.oc.df()
my.ss.bin.trt.oc.df
}
