% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ts.bin.studyend.GNG.r
\name{get.ts.bin.studyend.GNG}
\alias{get.ts.bin.studyend.GNG}
\title{Get Two-sample binary study end GNG}
\usage{
get.ts.bin.studyend.GNG(
  a.con = 1,
  b.con = 1,
  n.con = 30,
  x.con = 18,
  a.trt = 1,
  b.trt = 1,
  n.trt = 30,
  x.trt = 14,
  Delta.lrv = 0.3,
  Delta.tv = 0.4,
  tau.tv = 0.1,
  tau.lrv = 0.8,
  tau.ng = 0.65
)
}
\arguments{
\item{a.con}{prior alpha parameter for control group}

\item{b.con}{prior beta parameter for control group}

\item{n.con}{number of trials for control group}

\item{x.con}{number of responses for control group}

\item{a.trt}{prior alpha parameter for treatment group}

\item{b.trt}{prior beta parameter for treatment group}

\item{n.trt}{number of trials for treatment group}

\item{x.trt}{number of responses for treatment group}

\item{Delta.lrv}{TPP Lower Reference Value aka Min TPP}

\item{Delta.tv}{TPP Target Value aka Base TPP}

\item{tau.tv}{threshold associated with Base TPP}

\item{tau.lrv}{threshold associated with Min TPP}

\item{tau.ng}{threshold associated with No-Go}
}
\value{
returns a list holding data.frames alterting to what is needed from data to acheive go/no-go
}
\description{
Get Two-sample binary study end GNG
}
\examples{
get.ts.bin.studyend.GNG()
}
