% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tte.df.r
\name{get.tte.df}
\alias{get.tte.df}
\title{Get time to event data.frame}
\usage{
get.tte.df(
  m.con.prior = 50,
  m.trt.prior = 50,
  HR.prior = 0.845,
  HR.obs = seq(0.3, 1, 0.01),
  m.obs = seq(10, 200, 5),
  ARatio = 0.5,
  HR.tv = 0.8,
  HR.lrv = 0.9,
  tau.tv = 0.1,
  tau.lrv = 0.2,
  tau.ng = 0.35
)
}
\arguments{
\item{m.con.prior}{number of prior events for control}

\item{m.trt.prior}{number of prior events for treamtent}

\item{HR.prior}{HR estimate}

\item{HR.obs}{Observed HR}

\item{m.obs}{Obeserved number of events}

\item{ARatio}{Randomization ratio}

\item{HR.tv}{Base TPP for HR}

\item{HR.lrv}{Min TPP for HR}

\item{tau.tv}{threshold associated with Base TPP}

\item{tau.lrv}{threshold associated with Min TPP}

\item{tau.ng}{threshold associated with No-Go}
}
\value{
A data.frame is returned
}
\description{
Get time to event data.frame
}
\examples{
my.tte.df <- get.tte.df()
head(my.tte.df)
}
