% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tte.post.param.r
\name{get.tte.post.param}
\alias{get.tte.post.param}
\title{Get TTE posterior parameters}
\usage{
get.tte.post.param(
  m.con.prior = 10,
  m.trt.prior = 10,
  HR.prior = 0.7,
  ARatio = 1,
  HR.obs = 0.8,
  m.obs = 50
)
}
\arguments{
\item{m.con.prior}{prior number of events for control group}

\item{m.trt.prior}{prior number of events for treatment group}

\item{HR.prior}{hazard ratio estimate}

\item{ARatio}{randomization ratio}

\item{HR.obs}{HR observed}

\item{m.obs}{number of events}
}
\value{
A data.frame is returned.
}
\description{
Update parameters in time to event case
}
\examples{
my.tte.post.param <- get.tte.post.param()
my.tte.post.param
}
