% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tte.post.r
\name{get.tte.post}
\alias{get.tte.post}
\title{Get posterior parameters for two-sample time to event case using normal approximation to log(HR) - expand.grid}
\usage{
get.tte.post(
  m.con.prior = 10,
  m.trt.prior = 10,
  HR.prior = 0.7,
  ARatio = 1,
  HR.obs = 0.8,
  m.obs = 50
)
}
\arguments{
\item{m.con.prior}{prior number of events on control}

\item{m.trt.prior}{prior number of events on treatment}

\item{HR.prior}{Prior Hazard ratio}

\item{ARatio}{Randomization ratio in current trial}

\item{HR.obs}{Observed HR in current trial}

\item{m.obs}{Total events in current trail}
}
\value{
A data.frame is returned
}
\description{
Returns the parameters of the posterior distribution in the case of normal approximation to hazard ratio
}
\examples{
my.tte.post <- get.tte.post()
my.tte.post
}
