% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tte.ssize.oc.df.r
\name{get.tte.ssize.oc.df}
\alias{get.tte.ssize.oc.df}
\title{Get time to event sample size OC data.frame}
\usage{
get.tte.ssize.oc.df(
  m.con.prior = 10,
  m.trt.prior = 10,
  HR.prior = 0.75,
  ARatio = 1,
  m.obs = 50,
  m.lower = 40,
  m.upper = 120,
  HR.lrv = 0.75,
  HR.tv = 0.75,
  HR.user = 0.845,
  tau.tv = 0.1,
  tau.lrv = 0.8,
  tau.ng = 0.65
)
}
\arguments{
\item{m.con.prior}{number of prior events for control}

\item{m.trt.prior}{number of prior events for treamtent}

\item{HR.prior}{HR estimate}

\item{ARatio}{randomization ratio}

\item{m.obs}{observed number of events}

\item{m.lower}{lower bound on number of events}

\item{m.upper}{upper bound on number of events}

\item{HR.lrv}{Min TPP for HR}

\item{HR.tv}{Base TPP for HR}

\item{HR.user}{user's HR}

\item{tau.tv}{threshold associated with Base TPP}

\item{tau.lrv}{threshold associated with Min TPP}

\item{tau.ng}{threshold associated with No-Go}
}
\value{
A data frame is returned
}
\description{
Get time to event sample size OC data.frame
}
\examples{
\donttest{
my.tte.ssize.oc.df <- get.tte.ssize.oc.df()
my.tte.ssize.oc.df
}
}
