% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tte.trt.oc.df.r
\name{get.tte.trt.oc.df}
\alias{get.tte.trt.oc.df}
\title{Get TTE treatment effect OC curve data.frame}
\usage{
get.tte.trt.oc.df(
  m.con.prior = 10,
  m.trt.prior = 10,
  HR.prior = 0.8,
  ARatio = 0.5,
  m.obs = 50,
  HR.tv = 0.7,
  HR.lrv = 0.9,
  HR.lower = 0.3,
  HR.upper = 2,
  tau.tv = 0.1,
  tau.lrv = 0.8,
  tau.ng = 0.65
)
}
\arguments{
\item{m.con.prior}{number of prior events for control}

\item{m.trt.prior}{number of prior events for treamtent}

\item{HR.prior}{HR estimate}

\item{ARatio}{randomization ratio}

\item{m.obs}{observed events}

\item{HR.tv}{HR for Base TPP}

\item{HR.lrv}{HR for Min TPP}

\item{HR.lower}{Lower bound for OC curve}

\item{HR.upper}{Upper bound for OC curve}

\item{tau.tv}{threshold associated with Base TPP}

\item{tau.lrv}{threshold associated with Min TPP}

\item{tau.ng}{threshold associated with No-Go}
}
\value{
A data.frame is returned
}
\description{
Get time to event treatment oc data.frame
}
\examples{
my.tte.trt.oc.df <- get.tte.trt.oc.df()
my.tte.trt.oc.df
}
\author{
Greg Cicconetti
}
