% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goparallel.R
\name{goparallel}
\alias{goparallel}
\title{Goparallel - initiate a parallel computing environment}
\usage{
goparallel(ncores = 7)
}
\arguments{
\item{ncores}{number of cores}
}
\value{
This function initiates a parallel computing environment based on the parallel package
}
\description{
Goparallel - initiate a parallel computing environment
}
