% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.bin.int.HelperFunctions.r
\name{Intr_func}
\alias{Intr_func}
\title{Intr_func}
\usage{
Intr_func(M, R, ORR)
}
\arguments{
\item{M}{M}

\item{R}{R}

\item{ORR}{ORR}
}
\value{
These functions are used by other functions
}
\description{
Intr_func
}
