% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_ls.r
\name{Location.scale.t}
\alias{Location.scale.t}
\alias{dt_ls}
\alias{pt_ls}
\alias{qt_ls}
\alias{rt_ls}
\title{Location-scale t distribution functions}
\usage{
dt_ls(x, df, mu, sigma)

pt_ls(x, df, mu, sigma)

qt_ls(prob, df, mu, sigma)

rt_ls(n, df, mu, sigma)
}
\arguments{
\item{x}{likelihood function evaluates at point x}

\item{df}{degrees of freedom}

\item{mu}{mean}

\item{sigma}{scale parameter}

\item{prob}{cumulative probability}

\item{n}{sample size}
}
\value{
Returns density values, cumulative probabilities, quantiles and random samples from a Location-scale t distribution.
}
\description{
Location-scale t distribution functions
}
\examples{
{
dt_ls(0, 100, 0, 1)
pt_ls(0, 100, 0, 1)
qt_ls(0.5, 100, 0, 1)
rt_ls(100, 100, 0, 1)
}
}
\author{
Greg Cicconetti
}
