% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.int.dec.GridPlot.r
\name{make.ss.bin.int.dec.GridPlot}
\alias{make.ss.bin.int.dec.GridPlot}
\title{Make single sample binary interim decision grid}
\usage{
make.ss.bin.int.dec.GridPlot(
  InterimDF,
  goThreshold = 0.8,
  nogoThreshold = 0.8,
  include_nogo = TRUE,
  lower.bound.go = 5,
  lower.bound.ng = 10,
  add.black = FALSE
)
}
\arguments{
\item{InterimDF}{call to get.ss.bin.interim.df}

\item{goThreshold}{predictive probability threshold}

\item{nogoThreshold}{predictive probability threshold}

\item{include_nogo}{logical}

\item{lower.bound.go}{lower bound for Go}

\item{lower.bound.ng}{Lower (upper!) bound for no-go}

\item{add.black}{logical}
}
\value{
A ggplot object is returned.
}
\description{
Make single sample binary interim decision grid
}
