% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.int.GStable.r
\name{make.ss.bin.int.GStable}
\alias{make.ss.bin.int.GStable}
\title{Make single sample binary interim Group-sequential style table}
\usage{
make.ss.bin.int.GStable(InterimDF, goThreshold = 0.8, nogoThreshold = 0.8)
}
\arguments{
\item{InterimDF}{call to get.ss.bin.interim.df}

\item{goThreshold}{predictive probability threshold}

\item{nogoThreshold}{predictive probability threshold}
}
\value{
A data.frame is returned
}
\description{
Make single sample binary interim Group-sequential style table
}
