% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.int.oc.table.r
\name{make.ss.bin.int.oc.table}
\alias{make.ss.bin.int.oc.table}
\title{Make single sample binary interim OC table}
\usage{
make.ss.bin.int.oc.table(Sims, assessment = "all", rate = 0.3)
}
\arguments{
\item{Sims}{call to get.ss.bin.int.oc.sim}

\item{assessment}{assessment}

\item{rate}{rate}
}
\value{
A data.frame is returned
}
\description{
Make single sample binary interim OC table
}
