% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.int.req.plot.r
\name{make.ss.bin.int.req.plot}
\alias{make.ss.bin.int.req.plot}
\title{Make single sample binary predictive probability plot}
\usage{
make.ss.bin.int.req.plot(my.table = make.ss.bin.int.req.table())
}
\arguments{
\item{my.table}{output from make.ss.bin.pp.table}
}
\value{
A ggplot object is returned.
}
\description{
Make single sample binary predictive probability plot
}
\examples{
my.ss.bin.int.data.req.plot<- make.ss.bin.int.req.plot()
my.ss.bin.int.data.req.plot
}
