% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.int.req.plot2.r
\name{make.ss.bin.int.req.plot2}
\alias{make.ss.bin.int.req.plot2}
\title{Make single sample binary interim Group-sequential style plot}
\usage{
make.ss.bin.int.req.plot2(
  InterimDF,
  Delta.lrv = 0.2,
  Delta.tv = 0.3,
  tsize = 5
)
}
\arguments{
\item{InterimDF}{call to get.ss.bin.interim.df}

\item{Delta.lrv}{min TPP}

\item{Delta.tv}{Base TPP}

\item{tsize}{text size - default is for app}
}
\value{
A ggplot object is returned
}
\description{
Make single sample binary interim Group-sequential style plot
}
