% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.int.req.table.r
\name{make.ss.bin.int.req.table}
\alias{make.ss.bin.int.req.table}
\title{Make single sample binary pred prob table}
\usage{
make.ss.bin.int.req.table(my.df = return.ss.bin.int.req())
}
\arguments{
\item{my.df}{output from return.ss.bin.int.data.req}
}
\value{
A data.frame is returned
}
\description{
Make single sample binary pred prob table
}
\examples{
my.ss.bin.int.req.table <- make.ss.bin.int.req.table()
head(my.ss.bin.int.req.table)
}
