% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.ppp.r
\name{make.ss.bin.ppp}
\alias{make.ss.bin.ppp}
\title{Make single sample binary prior/posterior plot}
\usage{
make.ss.bin.ppp(a.trt = 1, b.trt = 1, n.trt = 40, x.trt = 20)
}
\arguments{
\item{a.trt}{prior alpha parameter}

\item{b.trt}{prior beta parameter}

\item{n.trt}{number of trials}

\item{x.trt}{number of responses}
}
\value{
A ggplot object is returned.
}
\description{
Make single sample binary prior/posterior plot
}
\examples{
my.ss.bin.ppp <- make.ss.bin.ppp(a.trt = 1, b.trt = 1, n.trt = 40, x.trt = 20)
my.ss.bin.ppp[[1]]
my.ss.bin.ppp[[2]]
}
\author{
Greg Cicconetti
}
