% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.ssize.oc.r
\name{make.ss.bin.ssize.oc}
\alias{make.ss.bin.ssize.oc}
\title{Make single sample binary sample size OC curve}
\usage{
make.ss.bin.ssize.oc(
  for.plot = get.ss.bin.ssize.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{ouput from get.ss.bin.ssize.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned
}
\description{
Make single sample binary sample size OC curve
}
\examples{
my.ss.bin.ssize.oc.df <- get.ss.bin.ssize.oc.df()
my.ss.bin.ssize.oc <- make.ss.bin.ssize.oc(for.plot= my.ss.bin.ssize.oc.df, add.table=TRUE)
plot(my.ss.bin.ssize.oc)
}
