% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.ssize.oc2.r
\name{make.ss.bin.ssize.oc2}
\alias{make.ss.bin.ssize.oc2}
\title{Make single sample binary sample size OC - trend version}
\usage{
make.ss.bin.ssize.oc2(
  for.plot = get.ss.bin.ssize.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{output from get.ss.bin.ssize.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{provides extended output summaries}
}
\value{
A list is returned
}
\description{
Make single sample binary sample size OC - trend version
}
\examples{
my.ss.bin.ssize.oc.df <-get.ss.bin.ssize.oc.df()
my.ss.bin.ssize.oc2 <- make.ss.bin.ssize.oc2(for.plot = my.ss.bin.ssize.oc.df, add.table=TRUE)
plot(my.ss.bin.ssize.oc2)
}
\author{
Greg Cicconetti
}
