% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.trt.int.oc.r
\name{make.ss.bin.trt.int.oc}
\alias{make.ss.bin.trt.int.oc}
\title{Make single sample binary Operating characteristic function vs. treatment effect}
\usage{
make.ss.bin.trt.int.oc(results = get.ss.bin.trt.int.oc.df())
}
\arguments{
\item{results}{output from get.ss.bin.trt.int.oc.df}
}
\value{
A ggplot object is returned
}
\description{
Make single sample binary Operating characteristic function vs. treatment effect
}
\examples{
\donttest{
make.ss.bin.trt.int.oc()
}
}
\author{
Greg Cicconetti
}
