% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.bin.trt.oc2.r
\name{make.ss.bin.trt.oc2}
\alias{make.ss.bin.trt.oc2}
\title{Make single sample binary treatment oc curve version 2}
\usage{
make.ss.bin.trt.oc2(
  my.df = get.ss.bin.trt.oc.df(),
  tsize = 4,
  nlines = 25,
  nlines.ria = 20,
  Delta_OC_LB = 0,
  Delta_OC_UB = 1,
  add.table = TRUE
)
}
\arguments{
\item{my.df}{output from get.ss.bin.trt.oc.df}

\item{tsize}{Control for text size}

\item{nlines}{Control for text spacing}

\item{nlines.ria}{Control for text spacing}

\item{Delta_OC_LB}{Lower bound for OC curve}

\item{Delta_OC_UB}{Upper bound for OC curve}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned.
}
\description{
Make single sample binary treatment oc curve version 2
}
\examples{
make.ss.bin.trt.oc2()
}
