% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.ng.int.req.plot.r
\name{make.ss.ng.int.req.plot}
\alias{make.ss.ng.int.req.plot}
\title{Make single sample normal-gamma interim data requirement plot}
\usage{
make.ss.ng.int.req.plot(my.table = make.ss.ng.int.req.table())
}
\arguments{
\item{my.table}{output from make.ss.ng.pp.table}
}
\value{
A ggplot object is returned
}
\description{
Make single sample normal-gamma interim data requirement plot
}
\examples{
my.ss.ng.int.req.table <- make.ss.ng.int.req.table()
my.ss.ng.int.req.plot <- make.ss.ng.int.req.plot(my.table = my.ss.ng.int.req.table)
my.ss.ng.int.req.plot
}
