% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.ng.int.req.table.r
\name{make.ss.ng.int.req.table}
\alias{make.ss.ng.int.req.table}
\title{Make single sample normal-gamma interim data requirements table}
\usage{
make.ss.ng.int.req.table(my.params = return.ss.ng.int.req())
}
\arguments{
\item{my.params}{output from return.ss.ng.int.req}
}
\value{
A data.frame is returned
}
\description{
Make single sample normal-gamma interim data requirements table
}
\examples{
my.ss.ng.int.req.table <- make.ss.ng.int.req.table()
head(my.ss.ng.int.req.table)
}
