% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.ng.ssize.oc.r
\name{make.ss.ng.ssize.oc}
\alias{make.ss.ng.ssize.oc}
\title{Make single sample normal gamma sample size OC}
\usage{
make.ss.ng.ssize.oc(
  for.plot = get.ss.ng.ssize.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{output from get.ss.ng.ssize.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{use TRUE for verbose output}
}
\value{
a list is returned
}
\description{
Make single sample normal gamma sample size OC
}
\examples{
\donttest{
my.ss.ng.ssize.oc.df <- get.ss.ng.ssize.oc.df()
my.ss.ng.ssize.oc <- make.ss.ng.ssize.oc(for.plot=my.ss.ng.ssize.oc.df, add.table=TRUE)
plot(my.ss.ng.ssize.oc)
}
}
