% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.ng.trt.int.oc.r
\name{make.ss.ng.trt.int.oc}
\alias{make.ss.ng.trt.int.oc}
\title{Make single sample normal-gamma interim treatment oc curve}
\usage{
make.ss.ng.trt.int.oc(
  my.df = get.ss.ng.trt.int.oc.df(npoints = 20, n.MC = 1000, include_nogo = FALSE),
  include_nogo = FALSE
)
}
\arguments{
\item{my.df}{output from get.ss.ng.trt.int.oc.df}

\item{include_nogo}{logical}
}
\value{
A ggplot object is returned.
}
\description{
Make single sample normal-gamma interim treatment oc curve
}
\examples{
\donttest{
my.ss.ng.trt.int.oc.df <- get.ss.ng.trt.int.oc.df(npoints = 20, n.MC = 1000,
include_nogo = FALSE)
my.ss.ng.trt.int.oc <- make.ss.ng.trt.int.oc(my.df = my.ss.ng.trt.int.oc.df,
include_nogo=FALSE)
my.ss.ng.trt.int.oc[[1]]
my.ss.ng.trt.int.oc[[2]]
my.ss.ng.trt.int.oc[[3]]
}
}
