% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.ng.trt.oc1.r
\name{make.ss.ng.trt.oc1}
\alias{make.ss.ng.trt.oc1}
\title{Make single sample normal-gamma treatment OC curve}
\usage{
make.ss.ng.trt.oc1(
  my.df = get.ss.ng.trt.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{my.df}{output from get.ss.ng.trt.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned
}
\description{
Make single sample normal-gamma treatment OC curve
}
\examples{
my.ss.ng.trt.oc.df <- get.ss.ng.trt.oc.df()
my.ss.ng.trt.oc1 <- make.ss.ng.trt.oc1(my.df =my.ss.ng.trt.oc.df, add.table=TRUE)
plot(my.ss.ng.trt.oc1[[1]])
}
