% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ss.ng.trt.oc2.r
\name{make.ss.ng.trt.oc2}
\alias{make.ss.ng.trt.oc2}
\title{Make single sample normal-gamma treatment OC curve V2}
\usage{
make.ss.ng.trt.oc2(
  my.df = get.ss.ng.trt.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{my.df}{output from get.ss.ng.trt.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned.
}
\description{
Make single sample normal-gamma treatment OC curve V2
}
\examples{
\donttest{
make.ss.ng.trt.oc2()
}
}
\author{
Greg Cicconetti
}
