% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.int.dec.df.r
\name{make.ts.bin.int.dec.df}
\alias{make.ts.bin.int.dec.df}
\title{Make two-sample binary interim decision table}
\usage{
make.ts.bin.int.dec.df(DecisionTable)
}
\arguments{
\item{DecisionTable}{the results from get.ts.bin.int.dec}
}
\value{
a simplified table showing the go/nogo thresholds for each number  of control successes.
}
\description{
Make two-sample binary interim decision table
}
\examples{
my.ts.bin.int.dec <- get.ts.bin.int.dec()
my.ts.bin.int.dec.df <- make.ts.bin.int.dec.df(DecisionTable = my.ts.bin.int.dec)
head(my.ts.bin.int.dec.df)
}
