% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.int.dec.plot.r
\name{make.ts.bin.int.dec.plot}
\alias{make.ts.bin.int.dec.plot}
\title{Make two-sample binary interim decision plot}
\usage{
make.ts.bin.int.dec.plot(for.plot)
}
\arguments{
\item{for.plot}{the results from get.ts.bin.decision}
}
\value{
a ggplot object showing the decision points for the interim analysis.
}
\description{
Make two-sample binary interim decision plot
}
\examples{
my.ts.bin.int.dec <- get.ts.bin.int.dec(
a.con = 1, b.con = 1, a.trt = 1, b.trt = 1,
Delta.lrv = 0.15, Delta.tv = .30,
tau.tv = 0.10, tau.lrv = 0.80, tau.ng = 0.65,
go.thresh=.8, ng.thresh=.8,
n.trt = 40, n.con = 40,
n.int.c = c(10, 20, 30), n.int.t = c(10, 20, 30),
runs=500, include_nogo = TRUE)
my.ts.bin.int.dec.plot <- make.ts.bin.int.dec.plot(for.plot=my.ts.bin.int.dec)
my.ts.bin.int.dec.plot
}
