% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.int.oc.plot.r
\name{make.ts.bin.int.oc.plot}
\alias{make.ts.bin.int.oc.plot}
\title{Make two-sample binary interim oc plot}
\usage{
make.ts.bin.int.oc.plot(
  ts.bin.int.oc = get.ts.bin.int.oc(a.con = 1, b.con = 1, a.trt = 1, b.trt = 1, Delta.tv
    = 0.3, Delta.lrv = 0.2, tau.tv = 0.1, tau.lrv = 0.8, tau.ng = 0.65, go.thresh = 0.8,
    ng.thresh = 0.8, n.con = 40, n.trt = 40, n.int.c = c(10, 20, 30), n.int.t = c(10, 20,
    30), DecisionTable = NULL, runs = 5000, ControlRate = 0.2, TreatmentEffect = seq(0,
    0.8, 0.1), include_nogo = TRUE),
  nlines = 25,
  tsize = 4,
  include_nogo = TRUE
)
}
\arguments{
\item{ts.bin.int.oc}{Results from get.ts.bin.int.oc}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{include_nogo}{logical}
}
\value{
A ggplot showing the probability of each out come given a particular difference from control
}
\description{
Make two-sample binary interim oc plot
}
