% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.int.oc.table.r
\name{make.ts.bin.int.oc.table}
\alias{make.ts.bin.int.oc.table}
\title{Make two-sample binary interim OC table}
\usage{
make.ts.bin.int.oc.table(
  OCsims = get.ts.bin.int.oc(),
  Diff = 0.1,
  interim = "any"
)
}
\arguments{
\item{OCsims}{results from get.ts.bin.int.oc}

\item{Diff}{the difference of interest for the table, must be a difference value included in OCsims}

\item{interim}{either a specific interim analysis or 'any'}
}
\value{
returns a 3x3 table showing the probabilities of each outcome for the interim and final analysis.
If interm is 'any' the interim analysis is the first non-consider result observed at any interim.
}
\description{
Make two-sample binary interim OC table
}
