% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.pp.heatmap.r
\name{make.ts.bin.pp.heatmap}
\alias{make.ts.bin.pp.heatmap}
\title{Make two-sample binary predictive probability heatmap}
\usage{
make.ts.bin.pp.heatmap(my.data = pp.table, go.thresh = 0.8, ng.thresh = 0.8)
}
\arguments{
\item{my.data}{output from return.ts.bin.pp.table}

\item{go.thresh}{go threshold}

\item{ng.thresh}{no-go threshold}
}
\value{
A ggplot object is returned.
}
\description{
Make two-sample binary predictive probability heatmap
}
\examples{
\donttest{
pp.table <- return.ts.bin.int.predprob.table(goparallel=FALSE)
make.ts.bin.pp.heatmap(my.data = pp.table, go.thresh=0.8, ng.thresh=0.8)
}
}
