% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.ppp.r
\name{make.ts.bin.ppp}
\alias{make.ts.bin.ppp}
\title{Make two-sample binary prior/posterior plot}
\usage{
make.ts.bin.ppp(
  a.con = 1,
  b.con = 1,
  n.con = 40,
  x.con = 9,
  a.trt = 1,
  b.trt = 1,
  n.trt = 40,
  x.trt = 17
)
}
\arguments{
\item{a.con}{prior alpha parameter for control group}

\item{b.con}{prior beta parameter for control group}

\item{n.con}{number of trials for control group}

\item{x.con}{number of responses for control group}

\item{a.trt}{prior alpha parameter for treatment group}

\item{b.trt}{prior alpha parameter for treatment group}

\item{n.trt}{number of trials for treatment group}

\item{x.trt}{number of responses for treatment group}
}
\value{
A list is returned.
}
\description{
Make two-sample binary prior/posterior plot
}
\examples{
my.ts.bin.ppp <- make.ts.bin.ppp()
my.ts.bin.ppp[[1]]
my.ts.bin.ppp[[2]]
}
