% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.ssize.oc.r
\name{make.ts.bin.ssize.oc}
\alias{make.ts.bin.ssize.oc}
\title{Make two-sample binary sample size operating characteristics curve}
\usage{
make.ts.bin.ssize.oc(
  for.plot = get.ts.bin.ssize.oc.df(),
  tsize = 4,
  nlines = 25,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{output from get.ts.bin.ssize.oc.df}

\item{tsize}{Control for text size}

\item{nlines}{Control for text spacing}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned
}
\description{
Make two-sample binary sample size operating characteristics curve
}
\examples{
\donttest{
my.ts.bin.ssize.oc.df <- get.ts.bin.ssize.oc.df()
make.ts.bin.ssize.oc(for.plot = my.ts.bin.ssize.oc.df)
}
}
