% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.studyend.GNG.hm.r
\name{make.ts.bin.studyend.GNG.hm}
\alias{make.ts.bin.studyend.GNG.hm}
\title{Make two-sample binary study end GNG heatmap}
\usage{
make.ts.bin.studyend.GNG.hm(my.df = return.ts.bin.studyend.GNG.hm.df())
}
\arguments{
\item{my.df}{output from return.ts.bin.studyend.GNG.hm.df}
}
\value{
A ggplot object is returned
}
\description{
Make two-sample binary study end GNG heatmap
}
\examples{
\donttest{
make.ts.bin.studyend.GNG.hm()
}
}
