% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.trt.oc1.r
\name{make.ts.bin.trt.oc1}
\alias{make.ts.bin.trt.oc1}
\title{Make two-sample binary treatment OC curve}
\usage{
make.ts.bin.trt.oc1(
  for.plot = get.ts.bin.trt.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{output from get.ts.bin.trt.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned.
}
\description{
Make two-sample binary treatment OC curve
}
\examples{
\donttest{
my.ts.bin.trt.oc.df <- get.ts.bin.trt.oc.df()
my.ts.bin.trt.oc1 <- make.ts.bin.trt.oc1(for.plot=my.ts.bin.trt.oc.df, add.table=TRUE)
plot(my.ts.bin.trt.oc1[[1]])
my.ts.bin.trt.oc1[[2]]
my.ts.bin.trt.oc1[[3]]
}
}
