% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.bin.trt.oc2.r
\name{make.ts.bin.trt.oc2}
\alias{make.ts.bin.trt.oc2}
\title{Make two-sample binary treatment OC curve V2}
\usage{
make.ts.bin.trt.oc2(
  for.plot = get.ts.bin.trt.oc.df(),
  tsize = 4,
  nlines = 25,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{output from get.ts.bin.trt.oc.df}

\item{tsize}{Control for text size}

\item{nlines}{Control for text spacing}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned
}
\description{
Make two-sample binary treatment OC curve V2
}
\examples{
\donttest{
my.ts.bin.trt.oc.df <- get.ts.bin.trt.oc.df()
my.ts.bin.trt.oc2 <- make.ts.bin.trt.oc2(for.plot=my.ts.bin.trt.oc.df, add.table=TRUE)
plot(my.ts.bin.trt.oc2[[1]])
my.ts.bin.trt.oc2[[2]]
my.ts.bin.trt.oc2[[3]]
}
}
