% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.ng.ssize.oc.r
\name{make.ts.ng.ssize.oc}
\alias{make.ts.ng.ssize.oc}
\title{Make two-sample normal gamma sample size OC curve}
\usage{
make.ts.ng.ssize.oc(
  for.plot = get.ts.ng.ssize.oc.df(),
  tsize = 4,
  nlines = 25,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{output from get.ts.ng.ssize.oc.df}

\item{tsize}{Control for text size}

\item{nlines}{Control for text spacing}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned
}
\description{
Make two-sample normal gamma sample size OC curve
}
\examples{
\donttest{
my.ts.ng.ssize.oc.df <- get.ts.ng.ssize.oc.df(goparallel=FALSE)
my.ts.ng.ssize.oc <- make.ts.ng.ssize.oc()
my.ts.ng.ssize.oc
}
}
