% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.ng.studyend.GNG.hm.r
\name{make.ts.ng.studyend.GNG.hm}
\alias{make.ts.ng.studyend.GNG.hm}
\title{Make two-sample normal-gamma studyend GNG Heatmap}
\usage{
make.ts.ng.studyend.GNG.hm(my.df = return.ts.ng.studyend.GNG.hm.df())
}
\arguments{
\item{my.df}{output from return.ts.ng.studyend.GNG.hm.df}
}
\value{
A ggplot object is returned
}
\description{
Make two-sample normal-gamma studyend GNG Heatmap
}
\examples{
\donttest{
make.ts.ng.studyend.GNG.hm()
}
}
\author{
Greg Cicconetti
}
