% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ts.ng.trt.oc2.r
\name{make.ts.ng.trt.oc2}
\alias{make.ts.ng.trt.oc2}
\title{Make two-sample normal-gamma treatment OC curve V2}
\usage{
make.ts.ng.trt.oc2(
  for.plot = get.ts.ng.trt.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{for.plot}{output from get.ts.ng.trt.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned
}
\description{
Make two-sample normal-gamma treatment OC curve V2
}
\examples{
\donttest{
my.ts.ng.trt.oc.df <- get.ts.ng.trt.oc.df(goparallel=FALSE)
make.ts.ng.trt.oc2(for.plot=my.ts.ng.trt.oc.df, add.table=TRUE)
}
}
