% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.tte.int.data.table.r
\name{make.tte.int.data.table}
\alias{make.tte.int.data.table}
\title{Make TTE predictive probability table}
\usage{
make.tte.int.data.table(my.params = return.tte.int.data.req())
}
\arguments{
\item{my.params}{output from return.tte.int.data.req}
}
\value{
A data.frame is returned
}
\description{
Make TTE predictive probability table
}
\examples{
my.tte.int.data.req <- return.tte.int.data.req()
my.tte.int.data.table <- make.tte.int.data.table(my.params = my.tte.int.data.req)
my.tte.int.data.table
}
