% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.tte.int.oc.r
\name{make.tte.int.oc}
\alias{make.tte.int.oc}
\title{Make TTE interim treatment OC curve}
\usage{
make.tte.int.oc(
  my.df = get.tte.int.oc.df(npoints = 20, include_nogo = TRUE),
  nlines = 25,
  tsize = 6,
  HR.lower = 0.0025,
  HR.upper = 2,
  include_nogo = TRUE
)
}
\arguments{
\item{my.df}{output from get.tte.int.OC.df}

\item{nlines}{number of lines}

\item{tsize}{text size}

\item{HR.lower}{lower bound for HR}

\item{HR.upper}{upper bound for HR}

\item{include_nogo}{logical}
}
\value{
A ggplot object is returned
}
\description{
Make TTE interim treatment OC curve
}
\examples{
\donttest{
make.tte.int.oc()
}
}
