% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.tte.ppp.r
\name{make.tte.ppp}
\alias{make.tte.ppp}
\title{Make Time to event prior/posterior plot}
\usage{
make.tte.ppp(
  m.con.prior = 10,
  m.trt.prior = 10,
  HR.prior = 0.845,
  ARatio = 1,
  HR.obs = 0.75,
  m.obs = 50
)
}
\arguments{
\item{m.con.prior}{prior number of control events}

\item{m.trt.prior}{prior number of treatment events}

\item{HR.prior}{HR estimate}

\item{ARatio}{Randomization ratio}

\item{HR.obs}{HR observed}

\item{m.obs}{Number of observed events}
}
\value{
A ggplot object is returned
}
\description{
Make Time to event prior/posterior plot
}
\examples{
make.tte.ppp()
}
