% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.tte.trt.oc1.r
\name{make.tte.trt.oc1}
\alias{make.tte.trt.oc1}
\title{Make time to event treatment OC curve}
\usage{
make.tte.trt.oc1(
  plot.df = get.tte.trt.oc.df(),
  nlines = 25,
  tsize = 4,
  add.table = TRUE
)
}
\arguments{
\item{plot.df}{output from get.tte.trt.oc.df}

\item{nlines}{Control for text spacing}

\item{tsize}{Control for text size}

\item{add.table}{provides extended output summaries}
}
\value{
A ggplot object is returned
}
\description{
Make time to event treatment OC curve
}
\examples{
my.tte.trt.oc.df <- get.tte.trt.oc.df()
make.tte.trt.oc1(plot.df = my.tte.trt.oc.df, add.table=TRUE)
}
