% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.bin.int.HelperFunctions.r
\name{ocTable_multi}
\alias{ocTable_multi}
\title{Single sample binary helper functions}
\usage{
ocTable_multi(DesignTable, TargetRate)
}
\arguments{
\item{DesignTable}{DesignTable}

\item{TargetRate}{TargetRate}
}
\value{
These functions are used by other functions

These functions are used by other functions
}
\description{
Single sample binary helper functions
}
