% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2beta_diff.r
\name{p2beta_diff}
\alias{p2beta_diff}
\title{Probability for difference of x or more between two beta distributions}
\usage{
p2beta_diff(x, a1, b1, a2, b2)
}
\arguments{
\item{x}{value}

\item{a1}{alpha for first}

\item{b1}{beta for first}

\item{a2}{alpha for second}

\item{b2}{beta for second}
}
\value{
returns the probability of a difference of x or more between two beta distributions.
This is a simplification of the original P2beta function simplified for vectorization.
Therre is a lot of space here for improved run times.
}
\description{
Probability for difference of x or more between two beta distributions
}
\author{
Randall Henner
}
