% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2beta_diff.r
\name{p2beta_diff_Vector}
\alias{p2beta_diff_Vector}
\title{Probability for difference of x or more between two beta distributions}
\usage{
p2beta_diff_Vector(x, a1, b1, a2, b2)
}
\arguments{
\item{x}{value}

\item{a1}{alpha for first}

\item{b1}{beta for first}

\item{a2}{alpha for second}

\item{b2}{beta for second}
}
\value{
a function is returned
}
\description{
Probability for difference of x or more between two beta distributions
}
