% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbetabinom_c.R
\name{pbetabinom_c}
\alias{pbetabinom_c}
\title{Replacement for pbetanimom}
\usage{
pbetabinom_c(q, size, m, s)
}
\arguments{
\item{q}{vector of quarantines}

\item{size}{vector of totals}

\item{m}{vector of probabilities of success}

\item{s}{vector of over dispersion parameters}
}
\value{
This is a modification of the pbetanimom function from the rmutil package.
 The original function would throw an error for non-nonsensical values rather than returning 0 or 1.
}
\description{
Replacement for pbetanimom
}
