% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled.sd.r
\name{pooled.sd}
\alias{pooled.sd}
\title{Pooled standard deviation}
\usage{
pooled.sd(s = c(4, 5), n = c(14, 20))
}
\arguments{
\item{s}{expression to be plotted}

\item{n}{number of points to plot}
}
\value{
A vector holding the pooled standard deviation is returned.
}
\description{
Computes to the pooled standard deviation.
}
\examples{
pooled.sd()
}
\author{
Greg Cicconetti
}
