% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.bin.int.HelperFunctions.r
\name{probmat.func}
\alias{probmat.func}
\title{probmat.func}
\usage{
probmat.func(DesignTable, TargetRate)
}
\arguments{
\item{DesignTable}{DesignTable}

\item{TargetRate}{TargetRate}
}
\value{
These functions are used by other functions
}
\description{
probmat.func
}
