% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SverdlovFunctions.r
\name{SverdlovFunction}
\alias{SverdlovFunction}
\alias{f}
\alias{appel.hypgeom}
\alias{g}
\alias{gauss.hypgeom}
\alias{root}
\alias{r2beta}
\alias{d2beta}
\alias{p2beta}
\alias{q2beta}
\alias{ci2beta}
\title{Sverdlov's functions}
\usage{
f(a, b1, b2, c, x, y, t)

appel.hypgeom(a, b1, b2, c, x, y)

g(a, b, c, x, t)

gauss.hypgeom(a, b, c, x)

root(fun, left, right, tol)

r2beta(relation, n, a1, b1, a2, b2)

d2beta(relation, x, a1, b1, a2, b2)

p2beta(relation, approach, x, a1, b1, a2, b2, n = 1e+06)

q2beta(relation, a1, b1, a2, b2, alpha, tol = 10^(-5))

ci2beta(relation, method, a1, b1, a2, b2, alpha, left0, right0)
}
\arguments{
\item{a}{a}

\item{b1}{b1}

\item{b2}{b2}

\item{c}{c}

\item{x}{x}

\item{y}{y}

\item{t}{t}

\item{b}{b}

\item{fun}{fun}

\item{left}{left}

\item{right}{right}

\item{tol}{tol}

\item{relation}{relation}

\item{n}{n}

\item{a1}{a1}

\item{a2}{a2}

\item{approach}{approach}

\item{alpha}{alpha}

\item{method}{method}

\item{left0}{left0}

\item{right0}{right0}
}
\value{
Various functions are offered.
}
\description{
Sverdlov's f function
}
\references{
Sverdlov O, Ryeznik Y, Wu S. Exact Bayesian Inference Comparing Binomial Proportions, With Application to Proof-of-Concept Clinical Trials. Therapeutic Innovation & Regulatory Science. 2015;49(1):163-174. doi:10.1177/2168479014547420
}
